Added checks for guilds that don't have settings
When a member joined/left a guild, the bot would crash if it could not find a setting in the DB for that guild.

Now we first check if a setting exists, if not return false.

(Setting files should be created on a 'first message' basis, but in the event it's overlooked we will return false here.)